requirements.txt
# BSP v1.0 - dependencias
# Core (bsp_core.py) - math/parsing/processing
numpy>=1.21.0
scipy>=1.7.0
openpyxl>=3.0.9

# GUI desktop (estabilidade_gui.py) - PDFs, imagens, plots
matplotlib>=3.5.0
reportlab>=3.6.0
python-docx>=0.8.11
Pillow>=9.0.0

# Dashboard web (bsp_dashboard/) - FastAPI + uvicorn local
fastapi>=0.110
uvicorn>=0.27
python-multipart>=0.0.6
pydantic>=2.0

# Build (opcional) - PyInstaller para .exe/.dmg, PyArmor para ofuscacao
# Activar PyArmor com BSP_OBFUSCATE=1 antes de correr BUILD_*.bat/sh
pyinstaller>=6.0; extra == 'build'
pyarmor>=8.5;    extra == 'build'
